+++ packages/commonjs/test/fixtures/function/dynamic-require-absolute-import/main.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-dynamic-require, global-require */
 
 t.deepEqual(require('./sub/submodule'), {
-	moduleDirect: 'direct',
-	moduleNested: 'nested',
-	parentModule: 'parent'
+  moduleDirect: 'direct',
+  moduleNested: 'nested',
+  parentModule: 'parent'
 });