+++ packages/commonjs/test/fixtures/function/dynamic-require-es-entry/importer.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-dynamic-require, global-require */
 
-function takeModule (withName) {
-	return require('./' + withName);
+function takeModule(withName) {
+  return require(`./${withName}`);
 }
 
 module.exports = takeModule('submodule.js');