+++ packages/commonjs/test/fixtures/function/dynamic-require-relative-paths/main.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-dynamic-require, global-require */
 
 function takeModuleWithDelimiter(name, delimiter) {
-	return require('.' + delimiter + name.replace(/=/g, delimiter));
+  return require(`.${delimiter}${name.replace(/=/g, delimiter)}`);
 }
 
 t.is(takeModuleWithDelimiter('sub=submodule.js', '/'), 'submodule');