+++ packages/commonjs/test/fixtures/function/dynamic-require-resolve-index/sub/sub.js
@@ -1,10 +1,10 @@
 /* eslint-disable import/no-dynamic-require, global-require */
 
 function takeModule(name) {
-	return require(name);
+  return require(name);
 }
 
 module.exports = {
-	parent: takeModule('..'),
-	customModule: takeModule('custom-module')
+  parent: takeModule('..'),
+  customModule: takeModule('custom-module')
 };