+++ packages/typescript/src/options/normalize.ts
@@ -38,6 +38,9 @@ export function normalizeCompilerOptions(
     // Default to using source maps.
     // If the plugin user sets sourceMap to false we keep that option.
     compilerOptions.sourceMap = true;
+    // Using inlineSources to make sure typescript generate source content
+    // instead of source path.
+    compilerOptions.inlineSources = true;
     autoSetSourceMap = true;
   }