+++ packages/beep/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rollup/plugin-beep ChangeLog
 
+## v0.2.0
+
+_2020-04-12_
+
+### Features
+
+- feat: Add types (#230)
+
 ## 0.1.0
 
 - First Release