+++ packages/node-resolve/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rollup/plugin-node-resolve ChangeLog
 
+## v7.1.3
+
+_2020-04-12_
+
+### Bugfixes
+
+- fix: resolve symlinked entry point properly (#291)
+
+## v7.1.2
+
+_2020-04-12_
+
+### Updates
+
+- docs: fix url (#289)
+
 ## v7.1.1
 
 _2020-02-03_